requirements.txt
tensorflow==1.15.3
pillow